Isolated static methods to be able to use for cut-plots
